Separate fs and mem DBs for easier bundling
