+++ lib/fs/index.ts
@@ -1,0 +1,5 @@
+export { default as fsDB } from "./jsongo.fs";
+
+export * from "./JsongoFSCollection";
+
+export * from "./JsongoFSDB";
