--- lib/jsongo.mem.ts
@@ -1,3 +1,0 @@
-import { JsongoMemDB } from "./JsongoDB";
-
-export default () => new JsongoMemDB();
