+++ lib/mem/index.ts
@@ -1,0 +1,5 @@
+export { default as memDB } from "./jsongo.mem";
+
+export * from "./JsongoMemCollection";
+
+export * from "./JsongoMemDB";
