+++ lib/mem/jsongo.mem.ts
@@ -1,0 +1,5 @@
+import { JsongoMemDB } from "./JsongoMemDB";
+
+const db = () => new JsongoMemDB();
+
+export default db;
