+++ lib/shared/index.ts
@@ -1,0 +1,7 @@
+export * from "./JsongoCollection";
+
+export * from "./JsongoDB";
+
+export * from "./JsongoDBProxy";
+
+export * from "./JsongoIDType";
